integra_reportes/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.0
python-dotenv>=1.0.0
